refactor: delete records action
